Break components down into smaller files
